Broken test scripts

+++ check_install.sh
@@ -6,5 +6,9 @@ pip install --upgrade virtualenv
 virtualenv ${ENV}
 source ${ENV}/bin/activate
 pip install .
+if [ "$?" != "0" ]
+then
+    exit 1
+fi
 deactivate
 rm -r ${ENV}
